pyproject.toml
[project]
name = "jsontable"
version = "0.1.0"
description = "A SQL generator for JSON_TABLE functions introduced in PostgreSQL 17"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "types-psycopg2",
    "psycopg2-binary",
    "sqlparse",
    "hatchling",
    "isort"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["src/jsontable"]

[tool.hatch.build.targets.wheel]
packages = ["src/jsontable"]

[tool.hatch.metadata]
root = "src"